PayWise/api.PayWise/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-unfold==0.47.0
sqlparse==0.5.3
